Address comments, move DownloadConfigType to near-config-utils.
