+++ tools/chainsync-loadtest/Cargo.toml
@@ -31,6 +31,7 @@ tokio.workspace = true
 
 near-async.workspace = true
 near-chain-configs.workspace = true
+near-config-utils.workspace = true
 near-crypto.workspace = true
 near-primitives.workspace = true
 near-store.workspace = true
